Updated with @abjerner's comments.
