+++ src/Skybrud.Umbraco.Redirects/Services/IRedirectsService.cs
@@ -1,5 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Linq;
+
 using Microsoft.AspNetCore.Http;
 using Skybrud.Umbraco.Redirects.Models;
 using Skybrud.Umbraco.Redirects.Models.Options;
@@ -121,9 +123,31 @@ namespace Skybrud.Umbraco.Redirects.Services {
         /// <summary>
         /// Returns an array of all rode nodes configured in Umbraco.
         /// </summary>
-        /// <param name="user">An <see cref="IUser"/> object to determine access to root nodes.</param>
-        /// <returns>An array of <see cref="RedirectRootNode"/> representing the root nodes.</returns>
-        RedirectRootNode[] GetRootNodes(IUser user);
+        /// <param name="user">An <see cref="IUser"/> with potential root node access restrictions.</param>
+        /// <returns>An array of <see cref="RedirectRootNode"/> representing the root nodes the user has access to.</returns>
+        RedirectRootNode[] GetRootNodes(IUser user)
+        {
+	        var rootNodes = GetRootNodes();
+	        HashSet<int> rootNodeIds = new();
+
+	        if (user.StartContentIds != null)
+	        {
+		        foreach (var rootNodeId in user.StartContentIds)
+		        {
+			        rootNodeIds.Add(rootNodeId);
+		        }
+	        }
+
+	        foreach (var group in user.Groups)
+	        {
+		        if (group.StartContentId != null)
+		        {
+			        rootNodeIds.Add(group.StartContentId.Value);
+		        }
+	        }
+
+	        return rootNodes.Where(rootNode => rootNodeIds.Any(x => rootNode.Path.Contains(x))).ToArray();
+        }
 
         /// <summary>
         /// Returns an array of redirects where the destination matches the specified <paramref name="nodeType"/> and <paramref name="nodeId"/>.
